+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 
 ### 💥 Breaking Changes
 
+* Blueprint has been upgraded from version 5 to version 6.  Most apps will not need to change,
+ but see https://github.com/palantir/blueprint/wiki/Blueprint-6.0 for more details.  Also note that
+ any custom css overrides to blueprint will need to be updated to refer to `bp6` instead of `bp5`.
 * Renamed `LoadSupport.loadModel` to `LoadSupport.loadObserver` for clarity. This property is a
   `TaskObserver` instance, not a `HoistModel`. The getter methods `HoistModel.loadModel` and
   `HoistService.loadModel` remain as aliases but are now deprecated and scheduled for removal in
@@ -36,6 +39,7 @@
   `react-grid-layout` v2+ (not common).
 * `DashCanvasModel.containerPadding` is now applied to the `react-grid-layout` div by RGL, not to the
   hoist-react provided containing div.  This may affect printing layouts.
+
 
 ### 🐞 Bug Fixes
 
@@ -63,6 +67,8 @@
 ### 📚 Libraries
 * react-grid-layout `1.5.0 → 2.1.1`
 * hoist-core `35.0.0`
+* @blueprintjs/core: 5.10.5 -> 6.3.2 ,
+* @blueprintjs/datetime: 5.3.7 -> 6.0.6,
 
 ## 78.1.4 - 2025-12-05